// apps/fusion/project/ws/app/src/lib/routes/infy/routes/navigator/models/navigator.model.ts
export interface IGroup {
  lp_groupid: string
  lp_groupdesc: string
  lp_groupname: string
  group_member: IGroupMember[]
  lp_groupimage: string
  certification_mandatory: boolean
}

export interface IVariant {
  variant_id: string
  variant_image: string
  variant_name: string
  variant_description: string
  group?: IGroup[]
  // certification: ICertification[]
}

export interface IRole {
  role_name: string
  role_id: string
  role_image: string
  variants: IVariant[]
  roles_defined: boolean
  courses_available: boolean
  role_description: string
}

export interface INsoData {
  arm_id: number
  roles: IRole[]
  arm_name: string
}

export interface IOfferings {
  Accelerate: INsoData
  Innovate: INsoData
  Insight: INsoData
  Experience: INsoData
  Assure: INsoData
}